Add file descriptor capabilities.
This change extends the fd_getfile() function to take a bitmask of
capabilities that are required by the call. All (most?) of the callers
have been updated to use one of the predefined capabilities.

As I am only interested in the capabilities useful for CloudABI, I am
adding an additional capability, CAP_OTHER, that is used as a catch-all
for all of the functions that do not have separate capabilities yet.

The file descriptor code itself has not been adjusted yet, meaning that
the capabilities are not respected yet.